requirements.txt
numpy
pandas 
scipy
numba
plotly
clickhouse-driver
pyodbc
sqlalchemy
psycopg2-binary
pytz
python-dateutil
jupyterlab
databases[asyncpg]
dotenv
pytest
databricks
databricks-sql
databricks-sql-connector